requirements.txt
# Production dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Development/testing dependencies
pytest>=7.0.0
build>=1.0.0